read microphone by spi

+++ Core/Inc/i2smic_spi.h
@@ -1,3 +1,3 @@
 #include "main.h"
 
-void I2SMic_Receive(uint16_t *pData, uint16_t Size, uint32_t Timeout);
+void I2SMic_Receive(uint32_t *pData, uint16_t Size, uint32_t Timeout);
